Refactor: Masonry layout 컴포넌트 분리
